+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [unreleased][unreleased]
+
+### Changed
+
+-   Improved support for using TPUEmbedding under parameter server strategy.
+
 ## [0.7.0][2022-07-07]
 
 A number of changes to make factorized top-K metric computation more accurate